+++ src/Filter/FilterInterface.php
@@ -37,7 +37,7 @@ interface FilterInterface
     public function getFieldName() : string;
     /**
      * Filter faceted data
-     * @param array<int|string,array<int>> $facetedData
+     * @param array<int|string,array<int>|\SplFixedArray<int>> $facetedData
      * @param array<int,bool|int>|null $inputIdKeys - RecordId passed into keys of an array (performance issue)
      * @return array<int,bool> - results in keys
      */