+++ src/Index/FixedArrayIndex.php
@@ -3,7 +3,7 @@
  *
  * MIT License
  *
- * Copyright (C) 2021  Kirill Yegorov https://github.com/k-samuel
+ * Copyright (C) 2021-2022 Kirill Yegorov https://github.com/k-samuel
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -29,88 +29,17 @@ declare(strict_types=1);
 
 namespace KSamuel\FacetedSearch\Index;
 
-use KSamuel\FacetedSearch\Indexer\IndexerInterface;
-
 /**
  * Simple faceted index
  * @package KSamuel\FacetedSearch
  */
-class FixedArrayIndex implements IndexInterface
+class FixedArrayIndex extends ArrayIndex
 {
-    /**
-     * Index data
-     * @var array<int|string,array<int|string,array<int>|\SplFixedArray<int>>>
-     */
-    protected $data = [];
-    /**
-     * @var null|array<int,bool>
-     */
-    private $idMapCache = null;
-    /**
-     * @var array<IndexerInterface>
-     */
-    protected $indexers = [];
     /**
      * @var bool
      */
     protected $isCompact = false;
 
-    /**
-     * Add record to index
-     * @param int $recordId
-     * @param array<int|string,array<int,mixed>> $recordValues -  ['fieldName'=>'fieldValue','fieldName2'=>['val1','val2']]
-     * @return bool
-     */
-    public function addRecord(int $recordId, array $recordValues): bool
-    {
-        if($this->isCompact){
-            throw new \RuntimeException('Cannot add record in readonly mode. Use $index->writeMode()');
-        }
-
-        $this->resetLocalCache();
-
-        foreach ($recordValues as $fieldName => $values) {
-            if (!is_array($values)) {
-                $values = [$values];
-            }
-
-            $values = array_unique($values);
-
-            if (isset($this->indexers[$fieldName])) {
-                if (!isset($this->data[$fieldName])) {
-                    $this->data[$fieldName] = [];
-                }
-                if (!$this->indexers[$fieldName]->add($this->data[$fieldName], $recordId, $values)) {
-                    return false;
-                }
-            } else {
-                foreach ($values as $value) {
-                    if (is_bool($value)) {
-                        $value = (int)$value;
-                    }
-                    if (is_float($value)) {
-                        $value = (string)$value;
-                    }
-                    $this->data[$fieldName][$value][] = $recordId;
-                }
-            }
-        }
-        return true;
-    }
-
-    /**
-     * Get facet data.
-     * @return array<int|string,array<int|string,array<int>|\SplFixedArray<int>>>
-     */
-    public function getData(): array
-    {
-        return $this->data;
-    }
-
-    private function resetLocalCache(): void
-    {
-        $this->idMapCache = null;
-    }
     /**
      * Get index data. Can be used for storing it to DB
      * @return array<int|string,array<int|string,array<int>>>
@@ -123,7 +52,7 @@ class FixedArrayIndex implements IndexInterface
         /**
          * @var array<int|string,array<int|string,array<int>>>
          */
-         return $this->data;
+        return $this->data;
     }
 
     /**
@@ -136,32 +65,39 @@ class FixedArrayIndex implements IndexInterface
              * @var \SplFixedArray<int> $recordList
              */
             foreach ($value as &$recordList) {
-                if($recordList instanceof \SplFixedArray){
+                if ($recordList instanceof \SplFixedArray) {
                     $recordList = $recordList->toArray();
                 }
             }
+            unset($recordList);
         }
+        unset($value);
         $this->isCompact = false;
     }
 
-    public function commitChanges() : void
+    /**
+     * Apply index updates (convert into \SplFixedArray)
+     */
+    public function commitChanges(): void
     {
         foreach ($this->data as &$value) {
             /**
              * @var array<int>|\SplFixedArray<int> $recordList
              */
             foreach ($value as &$recordList) {
-                if(is_array($recordList)){
+                if (is_array($recordList)) {
                     $recordList = \SplFixedArray::fromArray($recordList);
                 }
             }
+            unset($recordList);
         }
+        unset($value);
+
         $this->isCompact = true;
     }
 
     /**
-     * Set index data. Can be used for restoring from DB
-     * @param array<int|string,array<int|string,array<int>>> $data
+     * @inheritDoc
      */
     public function setData(array $data): void
     {
@@ -171,82 +107,20 @@ class FixedArrayIndex implements IndexInterface
         $this->commitChanges();
     }
 
-    /**
-     * Get field data section from index
-     * @param string $fieldName
-     * @return array<int|string,array<int>|\SplFixedArray<int>>
-     */
-    public function getFieldData(string $fieldName): array
-    {
-        return $this->data[$fieldName] ?? [];
-    }
-
 
     /**
-     * Get all records from index
-     * @return array<int>
+     * @inheritDoc
+     * Performance patch SplFixedArray index access is faster than iteration
      */
-    public function getAllRecordId(): array
+    protected function getIntersectMapCount($a, array $b): int
     {
-        return array_keys($this->getAllREcordIdMap());
-    }
-
-    /**
-     * Get all records from index as map [$id1=>true,...]
-     * @return array<int,bool>
-     */
-    public function getAllRecordIdMap(): array
-    {
-        if ($this->idMapCache !== null) {
-            return $this->idMapCache;
-        }
-
-        $result = [];
-        foreach ($this->data as $values) {
-            foreach ($values as $list) {
-                foreach ($list as $v) {
-                    $result[$v] = true;
-                }
+        $intersectLen = 0;
+        $count = count($a);
+        for ($i = 0; $i < $count; $i++) {
+            if (isset($b[$a[$i]])) {
+                $intersectLen++;
             }
         }
-        /**
-         * @var array<int,bool> $result
-         */
-
-        $this->idMapCache = $result;
-        return $result;
-    }
-
-    /**
-     * Add specialized indexer for field
-     * @param string $fieldName
-     * @param IndexerInterface $indexer
-     */
-    public function addIndexer(string $fieldName, IndexerInterface $indexer): void
-    {
-        $this->indexers[$fieldName] = $indexer;
-    }
-
-    /**
-     * @param string $field
-     * @param mixed $value
-     * @return int
-     */
-    public function getRecordsCount(string $field, $value): int
-    {
-        if (!isset($this->data[$field][$value])) {
-            return 0;
-        }
-        return count($this->data[$field][$value]);
-    }
-
-    /**
-     * Check if field exists
-     * @param string $fieldName
-     * @return bool
-     */
-    public function hasField(string $fieldName): bool
-    {
-        return isset($this->data[$fieldName]);
+        return $intersectLen;
     }
 }